Initial commit of refactor
Created two new "base" classes called `model` and `model-collection`. These two classes provide a way for creating models that are not the typical Nylas API models (Event, Calendar, etc.) but will still require similar functionality. Also not all models are "RESTful", `Contact.Email` for example does not.
